feat: added schema tests & fixed mapping
